+++ src/call/server.rs
@@ -3,6 +3,7 @@
 use std::ffi::CStr;
 use std::pin::Pin;
 use std::sync::Arc;
+use std::time::Duration;
 use std::{result, slice};
 
 use crate::grpc_sys::{
@@ -30,8 +31,10 @@ use crate::server::{BoxHandler, RequestCallContext};
 use crate::task::{BatchFuture, CallTag, Executor, Kicker};
 use crate::CheckResult;
 
+/// A time point that an rpc or operation should finished before it.
+#[derive(Clone, Copy)]
 pub struct Deadline {
-    spec: gpr_timespec,
+    pub(crate) spec: gpr_timespec,
 }
 
 impl Deadline {
@@ -44,11 +47,26 @@ impl Deadline {
         }
     }
 
-    pub fn exceeded(&self) -> bool {
+    /// Checks if the deadline is exceeded.
+    pub fn exceeded(self) -> bool {
         unsafe {
             let now = grpc_sys::gpr_now(gpr_clock_type::GPR_CLOCK_REALTIME);
             grpc_sys::gpr_time_cmp(now, self.spec) >= 0
         }
+    }
+
+    pub(crate) fn spec(self) -> gpr_timespec {
+        self.spec
+    }
+}
+
+impl From<Duration> for Deadline {
+    /// Build a deadline from given duration.
+    ///
+    /// The deadline will be `now + duration`.
+    #[inline]
+    fn from(dur: Duration) -> Deadline {
+        Deadline::new(dur.into())
     }
 }
 
@@ -234,7 +252,7 @@ impl UnaryRequestContext {
             return execute(self.request, cq, reader, handler, checker);
         }
 
-        let status = RpcStatus::new(RpcStatusCode::INTERNAL, Some("No payload".to_owned()));
+        let status = RpcStatus::with_message(RpcStatusCode::INTERNAL, "No payload".to_owned());
         self.request.call(cq.clone()).abort(&status)
     }
 }
@@ -635,8 +653,8 @@ impl<'a> RpcContext<'a> {
         self.ctx.host()
     }
 
-    pub fn deadline(&self) -> &Deadline {
-        &self.deadline
+    pub fn deadline(&self) -> Deadline {
+        self.deadline
     }
 
     /// Get the initial metadata sent by client.
@@ -694,9 +712,9 @@ pub fn execute_unary<P, Q, F>(
     let request = match de(payload) {
         Ok(f) => f,
         Err(e) => {
-            let status = RpcStatus::new(
+            let status = RpcStatus::with_message(
                 RpcStatusCode::INTERNAL,
-                Some(format!("Failed to deserialize response message: {:?}", e)),
+                format!("Failed to deserialize response message: {:?}", e),
             );
             call.abort(&status);
             return;
@@ -740,9 +758,9 @@ pub fn execute_server_streaming<P, Q, F>(
     let request = match de(payload) {
         Ok(t) => t,
         Err(e) => {
-            let status = RpcStatus::new(
+            let status = RpcStatus::with_message(
                 RpcStatusCode::INTERNAL,
-                Some(format!("Failed to deserialize response message: {:?}", e)),
+                format!("Failed to deserialize response message: {:?}", e),
             );
             call.abort(&status);
             return;
@@ -777,7 +795,7 @@ pub fn execute_unimplemented(ctx: RequestContext, cq: CompletionQueue) {
     let ctx = ctx;
     let mut call = ctx.call(cq);
     accept_call!(call);
-    call.abort(&RpcStatus::new(RpcStatusCode::UNIMPLEMENTED, None))
+    call.abort(&RpcStatus::new(RpcStatusCode::UNIMPLEMENTED))
 }
 
 // Helper function to call handler.